remove api calls
